+++ gamma_dashboard/static/dashboard/js/app/src/components/dashboard/charts/index.jsx
@@ -1,8 +1,2 @@
 export { PointsDistributionChart } from './points-distribution-chart';
 export { ProgressChart } from './progress-chart';
-export {
-  chartTitleStyles,
-  chartSubtitleStyles,
-  chartTitleOptions,
-  legendOptions,
-} from './constants';