+++ gamma_dashboard/static/dashboard/js/app/src/components/dashboard/charts/points-distribution-chart/PointsDistributionChart.jsx
@@ -1,26 +1,35 @@
-import React from 'react';
+import React, { useRef } from 'react';
 import PropTypes from 'prop-types';
+import ReactECharts from 'echarts-for-react';
+import { breakpoints, useMediaQuery } from '@openedx/paragon';
 
 import { useTranslate } from '../../../../i18n/utils';
-import { ChartWithExport } from '../chart-with-export';
-import { getPointsDistributionChartConfig } from './config';
+import { useElementWidth } from '../hooks';
+import { getConfig } from './config';
 import { prepareEvents } from './utils';
 
 const PointsDistributionChart = ({ data }) => {
+  const chartRef = useRef(null);
+  const chartWidth = useElementWidth(chartRef);
   const events = prepareEvents(data);
+  const isSmall = useMediaQuery({ maxWidth: breakpoints.extraSmall.maxWidth });
 
   const messages = {
-    distributionSection: {
-      headingText: useTranslate('performance.points.distribution.section.heading.text'),
-      descriptionText: useTranslate('performance.points.distribution.section.description.text'),
+    headingText: useTranslate('performance.points.distribution.section.heading.text'),
+    descriptionText: useTranslate('performance.points.distribution.section.description.text'),
+    seriesPointsName: useTranslate('performance.points.series.name'),
+    controls: {
+      saveAsImage: useTranslate('performance.points.controls.saveAsImage.label'),
     },
-    performancePointsSeriesName: useTranslate('performance.points.series.name'),
   };
 
   return (
-    <ChartWithExport
-      options={getPointsDistributionChartConfig(events, messages)}
-    />
+    <div ref={chartRef}>
+      <ReactECharts
+        option={getConfig(events, messages, chartWidth, isSmall)}
+        style={{ height: '470px' }}
+      />
+    </div>
   );
 };