+++ gamma_dashboard/static/dashboard/js/app/src/components/dashboard/components/progress-badge/popover-content/ProgressBlock.jsx
@@ -8,7 +8,7 @@ const ProgressBlock = ({ progressValues }) => {
   const { count = 0, goal = 0, title: progressTitle = '' } = progressValues || {};
 
   return (
-    <ul className="pl-3 pr-3">
+    <ul className="pl-3 pr-3 mb-0">
       <li>
         <Badge className="ml-2 mr-2" variant="info">
           {`${Math.min(count, goal)}/${goal}`}